🚀 Refactor tests
